FHAC-473: Moved the Async EJB logic to a separate module
