+++ Product/Production/Deploy/pom.xml
@@ -115,6 +115,7 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-ear-plugin</artifactId>
                 <configuration>
+                    <applicationName>CONNECT</applicationName>
                     <modules>
                         <webModule>
                             <groupId>org.connectopensource</groupId>
@@ -575,6 +576,12 @@
                     <version>${project.parent.version}</version>
                     <type>pom</type>
                 </dependency>
+                <dependency>
+                    <groupId>org.connectopensource</groupId>
+                    <artifactId>ConnectEJBCore</artifactId>
+                    <version>${project.parent.version}</version>
+                    <type>ejb</type>
+                </dependency>
             </dependencies>
 
             <build>
@@ -596,6 +603,11 @@
                                     <contextRoot>/Adapter/CORE_X12DocumentSubmission/A_0</contextRoot>
                                     <excluded>${x12.excluded}</excluded>
                                 </webModule>
+                                <ejbModule>
+                                   <groupId>org.connectopensource</groupId>
+                                   <artifactId>ConnectEJBCore</artifactId>
+                                    <excluded>${x12.excluded}</excluded>
+                                </ejbModule>
                             </modules>
                         </configuration>
                     </plugin>
