+++ Product/Production/Services/AuditRepositoryCore/src/main/java/gov/hhs/fha/nhinc/auditrepository/AuditRepositoryLogger.java
@@ -26,6 +26,8 @@
  */
 package gov.hhs.fha.nhinc.auditrepository;
 
+import gov.hhs.fha.nhinc.auditrepository.nhinc.proxy.AuditRepositoryProxy;
+import gov.hhs.fha.nhinc.auditrepository.nhinc.proxy.AuditRepositoryProxyObjectFactory;
 import gov.hhs.fha.nhinc.common.auditlog.AdhocQueryMessageType;
 import gov.hhs.fha.nhinc.common.auditlog.AdhocQueryResponseMessageType;
 import gov.hhs.fha.nhinc.common.auditlog.DocRetrieveMessageType;
@@ -58,12 +60,14 @@ import gov.hhs.fha.nhinc.common.hiemauditlog.LogSubscribeResponseType;
 import gov.hhs.fha.nhinc.common.hiemauditlog.LogUnsubscribeResponseType;
 import gov.hhs.fha.nhinc.common.hiemauditlog.SubscribeResponseMessageType;
 import gov.hhs.fha.nhinc.common.hiemauditlog.UnsubscribeResponseMessageType;
+import gov.hhs.fha.nhinc.common.nhinccommon.AcknowledgementType;
 import gov.hhs.fha.nhinc.common.nhinccommon.AssertionType;
 import gov.hhs.fha.nhinc.common.nhinccommon.NhinTargetSystemType;
 import gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayProvideAndRegisterDocumentSetSecuredRequestType;
 import gov.hhs.fha.nhinc.common.nhinccommoninternalorch.NotifyRequestType;
 import gov.hhs.fha.nhinc.common.nhinccommoninternalorch.SubscribeRequestType;
 import gov.hhs.fha.nhinc.common.nhinccommoninternalorch.UnsubscribeRequestType;
+import gov.hhs.fha.nhinc.corex12.docsubmission.audit.transform.CORE_X12AuditDataTransform;
 import gov.hhs.fha.nhinc.nhinclib.NhincConstants;
 import gov.hhs.fha.nhinc.transform.audit.AdminDistTransforms;
 import gov.hhs.fha.nhinc.transform.audit.DocumentQueryTransforms;
@@ -77,9 +81,10 @@ import gov.hhs.fha.nhinc.transform.audit.XDRTransforms;
 import gov.hhs.healthit.nhin.DocQueryAcknowledgementType;
 import gov.hhs.healthit.nhin.XDRAcknowledgementType;
 import ihe.iti.xds_b._2007.ProvideAndRegisterDocumentSetRequestType;
+import java.util.Date;
+import java.util.Properties;
 import oasis.names.tc.ebxml_regrep.xsd.rs._3.RegistryResponseType;
 import oasis.names.tc.emergency.edxl.de._1.EDXLDistribution;
-
 import org.apache.log4j.Logger;
 import org.hl7.v3.MCCIIN000002UV01;
 import org.hl7.v3.PRPAIN201305UV02;
@@ -92,7 +97,7 @@ import org.hl7.v3.RespondingGatewayPRPAIN201306UV02ResponseType;
  *
  * @author Jon Hoppesch
  */
-public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLogger  {
+public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLogger {
 
     private static final Logger LOG = Logger.getLogger(AuditRepositoryLogger.class);
     private final PatientDiscoveryTransforms pdAuditTransformer = new PatientDiscoveryTransforms();
@@ -100,6 +105,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
     private final AdminDistTransforms adAuditTransformer = new AdminDistTransforms();
     private final DocumentQueryTransforms dqAuditTransforms = new DocumentQueryTransforms();
     private NotifyTransforms transformLib = new NotifyTransforms();
+    private final CORE_X12AuditDataTransform coreX12AuditDataTransform = new CORE_X12AuditDataTransform();
 
     /**
      * Constructor code for the AuditRepositoryLogger. This instantiates the object.
@@ -116,10 +122,10 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logNhinPatientDiscAck(MCCIIN000002UV01 message, AssertionType assertion,
-            String direction, String _interface) {
+        String direction, String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logNhinPatientDiscReq(...)");
         LogEventRequestType auditMsg = null;
-            auditMsg = pdAuditTransformer.transformAck2AuditMsg(message, assertion, direction, _interface);
+        auditMsg = pdAuditTransformer.transformAck2AuditMsg(message, assertion, direction, _interface);
         LOG.debug("Exiting AuditRepositoryLogger.logNhinPatientDiscReq(...)");
         return auditMsg;
     }
@@ -134,11 +140,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logNhinPatientDiscReq(PRPAIN201305UV02 message, AssertionType assertion,
-            String direction, String _type) {
+        String direction, String _type) {
         LOG.debug("Entering AuditRepositoryLogger.logNhinPatientDiscReq(...)");
         LogEventRequestType auditMsg = null;
-            auditMsg = pdAuditTransformer.transformNhinPRPAIN201305RequestToAuditMsg(message, assertion, direction,
-                    NhincConstants.AUDIT_LOG_NHIN_INTERFACE, _type, null);
+        auditMsg = pdAuditTransformer.transformNhinPRPAIN201305RequestToAuditMsg(message, assertion, direction,
+            NhincConstants.AUDIT_LOG_NHIN_INTERFACE, _type, null);
         LOG.debug("Exiting AuditRepositoryLogger.logNhinPatientDiscReq(...)");
         return auditMsg;
     }
@@ -153,11 +159,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logNhinPatientDiscResp(PRPAIN201306UV02 message, AssertionType assertion,
-            String direction, String _type) {
+        String direction, String _type) {
         LOG.debug("Entering AuditRepositoryLogger.logNhinPatientDiscResp(...)");
         LogEventRequestType auditMsg = null;
-            auditMsg = pdAuditTransformer.transformNhinPRPAIN201306ResponseToAuditMsg(message, assertion, direction,
-                    NhincConstants.AUDIT_LOG_NHIN_INTERFACE, _type);
+        auditMsg = pdAuditTransformer.transformNhinPRPAIN201306ResponseToAuditMsg(message, assertion, direction,
+            NhincConstants.AUDIT_LOG_NHIN_INTERFACE, _type);
         LOG.debug("Exiting AuditRepositoryLogger.logNhinPatientDiscResp(...)");
         return auditMsg;
     }
@@ -172,11 +178,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logAdapterPatientDiscReq(PRPAIN201305UV02 message, AssertionType assertion,
-            String direction, String _type) {
+        String direction, String _type) {
         LOG.debug("Entering AuditRepositoryLogger.logAdapterPatientDiscReq(...)");
         LogEventRequestType auditMsg = null;
-            auditMsg = pdAuditTransformer.transformAdapterPRPAIN201305RequestToAuditMsg(message, assertion, direction,
-                    NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE, _type, null);
+        auditMsg = pdAuditTransformer.transformAdapterPRPAIN201305RequestToAuditMsg(message, assertion, direction,
+            NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE, _type, null);
         LOG.debug("Exiting AuditRepositoryLogger.logAdapterPatientDiscReq(...)");
         return auditMsg;
     }
@@ -191,11 +197,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logAdapterPatientDiscResp(PRPAIN201306UV02 message, AssertionType assertion,
-            String direction, String _type) {
+        String direction, String _type) {
         LOG.debug("Entering AuditRepositoryLogger.logAdapterPatientDiscResp(...)");
         LogEventRequestType auditMsg = null;
-            auditMsg = pdAuditTransformer.transformNhinPRPAIN201306ResponseToAuditMsg(message, assertion, direction,
-                    NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE, _type);
+        auditMsg = pdAuditTransformer.transformNhinPRPAIN201306ResponseToAuditMsg(message, assertion, direction,
+            NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE, _type);
         LOG.debug("Exiting AuditRepositoryLogger.logAdapterPatientDiscResp(...)");
         return auditMsg;
     }
@@ -210,11 +216,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityPatientDiscReq(RespondingGatewayPRPAIN201305UV02RequestType message,
-            AssertionType assertion, String direction, String _type, String _process) {
+        AssertionType assertion, String direction, String _type, String _process) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityPatientDiscReq(...)");
         LogEventRequestType auditMsg = null;
-            auditMsg = pdAuditTransformer.transformEntityPRPAIN201305RequestToAuditMsg(message, assertion, direction,
-                    NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, _type, _process);
+        auditMsg = pdAuditTransformer.transformEntityPRPAIN201305RequestToAuditMsg(message, assertion, direction,
+            NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, _type, _process);
         LOG.debug("Exiting AuditRepositoryLogger.logEntityPatientDiscReq(...)");
         return auditMsg;
     }
@@ -229,11 +235,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityPatientDiscResp(RespondingGatewayPRPAIN201306UV02ResponseType message,
-            AssertionType assertion, String direction, String _type) {
+        AssertionType assertion, String direction, String _type) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityPatientDiscResp(...)");
         LogEventRequestType auditMsg = null;
-            auditMsg = pdAuditTransformer.transformEntityPRPAIN201306ResponseToAuditMsg(message, assertion, direction,
-                    NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, _type);
+        auditMsg = pdAuditTransformer.transformEntityPRPAIN201306ResponseToAuditMsg(message, assertion, direction,
+            NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, _type);
         LOG.debug("Exiting AuditRepositoryLogger.logEntityPatientDiscResp(...)");
         return auditMsg;
     }
@@ -248,11 +254,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityPatientDiscAsyncResp(RespondingGatewayPRPAIN201306UV02RequestType message,
-            AssertionType assertion, String direction, String _type) {
+        AssertionType assertion, String direction, String _type) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityPatientDiscAsyncResp(...)");
         LogEventRequestType auditMsg = null;
         auditMsg = pdAuditTransformer.transformEntityPRPAIN201306AsyncResponseToAuditMsg(message, assertion, direction,
-                NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, _type);
+            NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, _type);
         LOG.debug("Exiting AuditRepositoryLogger.logEntityPatientDiscAsyncResp(...)");
         return auditMsg;
     }
@@ -266,11 +272,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityXDRReq(
-            RespondingGatewayProvideAndRegisterDocumentSetSecuredRequestType message, AssertionType assertion,
-            String direction) {
+        RespondingGatewayProvideAndRegisterDocumentSetSecuredRequestType message, AssertionType assertion,
+        String direction) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityXDRReq(...)");
         LogEventRequestType auditMsg = xdrAuditTransformer.transformRequestToAuditMsg(message.getProvideAndRegisterDocumentSetRequest(),
-                assertion, null, direction, NhincConstants.AUDIT_LOG_ENTITY_INTERFACE);
+            assertion, null, direction, NhincConstants.AUDIT_LOG_ENTITY_INTERFACE);
         LOG.debug("Exiting AuditRepositoryLogger.logEntityXDRReq(...)");
         return auditMsg;
     }
@@ -284,10 +290,10 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logAdapterXDRResponse(RegistryResponseType response, AssertionType assertion,
-            String direction) {
+        String direction) {
         LOG.debug("Entering AuditRepositoryLogger.logAdapterXDRResponse(...)");
         LogEventRequestType auditMsg = xdrAuditTransformer.transformResponseToAuditMsg(response, assertion, null,
-                direction, NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE, true);
+            direction, NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE, true);
         LOG.debug("Exiting AuditRepositoryLogger.logAdapterXDRResponse(...)");
         return auditMsg;
     }
@@ -301,10 +307,10 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityXDRResponse(RegistryResponseType response, AssertionType assertion,
-            String direction) {
+        String direction) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityXDRResponse(...)");
         LogEventRequestType auditMsg = xdrAuditTransformer.transformResponseToAuditMsg(response, assertion, null,
-                direction, NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, true);
+            direction, NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, true);
         LOG.debug("Exiting AuditRepositoryLogger.logEntityXDRResponse(...)");
         return auditMsg;
     }
@@ -318,11 +324,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityXDRResponseRequest(
-            gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayProvideAndRegisterDocumentSetSecuredResponseRequestType response,
-            AssertionType assertion, String direction) {
+        gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayProvideAndRegisterDocumentSetSecuredResponseRequestType response,
+        AssertionType assertion, String direction) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityXDRResponse(...)");
         LogEventRequestType auditMsg = xdrAuditTransformer.transformRequestToAuditMsg(response, assertion, null,
-                direction, NhincConstants.AUDIT_LOG_ENTITY_INTERFACE);
+            direction, NhincConstants.AUDIT_LOG_ENTITY_INTERFACE);
         LOG.debug("Exiting AuditRepositoryLogger.logEntityXDRResponse(...)");
         return auditMsg;
     }
@@ -350,14 +356,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logAdhocQuery(AdhocQueryMessageType message, String direction, String _interface,
-            String responseCommunityId) {
+        String responseCommunityId) {
         LOG.debug("Entering AuditRepositoryLogger.logAdhocQuery(...)");
         LogEventRequestType auditMsg = null;
-            LogAdhocQueryRequestType logReqMsg = new LogAdhocQueryRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            auditMsg = dqAuditTransforms.transformDocQueryReq2AuditMsg(logReqMsg, responseCommunityId);
+        LogAdhocQueryRequestType logReqMsg = new LogAdhocQueryRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        auditMsg = dqAuditTransforms.transformDocQueryReq2AuditMsg(logReqMsg, responseCommunityId);
         LOG.debug("Exiting AuditRepositoryLogger.logAdhocQuery(...)");
         return auditMsg;
     }
@@ -371,7 +377,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logAdhocQueryResult(AdhocQueryResponseMessageType message, String direction,
-            String _interface) {
+        String _interface) {
         return this.logAdhocQueryResult(message, direction, _interface, null);
     }
 
@@ -385,14 +391,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logAdhocQueryResult(AdhocQueryResponseMessageType message, String direction,
-            String _interface, String requestCommunityID) {
+        String _interface, String requestCommunityID) {
         LOG.debug("Entering AuditRepositoryLogger.logAdhocQueryResult(...)");
         LogEventRequestType auditMsg = null;
-            LogAdhocQueryResultRequestType logReqMsg = new LogAdhocQueryResultRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            auditMsg = dqAuditTransforms.transformDocQueryResp2AuditMsg(logReqMsg, requestCommunityID);
+        LogAdhocQueryResultRequestType logReqMsg = new LogAdhocQueryResultRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        auditMsg = dqAuditTransforms.transformDocQueryResp2AuditMsg(logReqMsg, requestCommunityID);
         LOG.debug("Exiting AuditRepositoryLogger.logAdhocQueryResult(...)");
         return auditMsg;
     }
@@ -407,7 +413,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logAdhocQueryDeferredAck(DocQueryAcknowledgementType acknowledgement,
-            AssertionType assertion, String direction, String _interface) {
+        AssertionType assertion, String direction, String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logAdhocQueryResult(...)");
         return logAdhocQueryDeferredAck(acknowledgement, assertion, direction, _interface, null);
     }
@@ -423,11 +429,11 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logAdhocQueryDeferredAck(DocQueryAcknowledgementType acknowledgement,
-            AssertionType assertion, String direction, String _interface, String requestCommunityID) {
+        AssertionType assertion, String direction, String _interface, String requestCommunityID) {
         LOG.debug("Entering AuditRepositoryLogger.logAdhocQueryResult(...)");
         LogEventRequestType auditMsg = null;
-            auditMsg = dqAuditTransforms.transformAcknowledgementToAuditMsg(acknowledgement, assertion, direction,
-                    _interface, requestCommunityID);
+        auditMsg = dqAuditTransforms.transformAcknowledgementToAuditMsg(acknowledgement, assertion, direction,
+            _interface, requestCommunityID);
         LOG.debug("Exiting AuditRepositoryLogger.logAdhocQueryResult(...)");
         return auditMsg;
     }
@@ -445,14 +451,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      */
     @Override
     public LogEventRequestType logDocRetrieve(DocRetrieveMessageType message, String direction, String _interface,
-            String responseCommunityID) {
+        String responseCommunityID) {
         LOG.debug("Entering AuditRepositoryLogger.logDocRetrieve(...)");
         LogEventRequestType auditMsg = null;
-            LogDocRetrieveRequestType logReqMsg = new LogDocRetrieveRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            auditMsg = DocumentRetrieveTransforms.transformDocRetrieveReq2AuditMsg(logReqMsg, responseCommunityID);
+        LogDocRetrieveRequestType logReqMsg = new LogDocRetrieveRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        auditMsg = DocumentRetrieveTransforms.transformDocRetrieveReq2AuditMsg(logReqMsg, responseCommunityID);
         LOG.debug("Exiting AuditRepositoryLogger.logDocRetrieve(...)");
         return auditMsg;
     }
@@ -462,7 +468,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      */
     @Override
     public LogEventRequestType logDocRetrieveResult(DocRetrieveResponseMessageType message, String direction,
-            String _interface) {
+        String _interface) {
         return logDocRetrieveResult(message, direction, _interface, null);
     }
 
@@ -471,14 +477,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      */
     @Override
     public LogEventRequestType logDocRetrieveResult(DocRetrieveResponseMessageType message, String direction,
-            String _interface, String requestCommunityID) {
+        String _interface, String requestCommunityID) {
         LOG.debug("Entering AuditRepositoryLogger.logDocRetrieveResult(...)");
         LogEventRequestType auditMsg = null;
-            LogDocRetrieveResultRequestType logReqMsg = new LogDocRetrieveResultRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            auditMsg = DocumentRetrieveTransforms.transformDocRetrieveResp2AuditMsg(logReqMsg, requestCommunityID);
+        LogDocRetrieveResultRequestType logReqMsg = new LogDocRetrieveResultRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        auditMsg = DocumentRetrieveTransforms.transformDocRetrieveResp2AuditMsg(logReqMsg, requestCommunityID);
         LOG.debug("Exiting AuditRepositoryLogger.logDocRetrieveResult(...)");
         return auditMsg;
     }
@@ -492,14 +498,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logFindAuditEvents(FindAuditEventsMessageType message, String direction,
-            String _interface) {
+        String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logFindAuditEvents(...)");
         LogEventRequestType auditMsg = null;
-            LogFindAuditEventsRequestType logReqMsg = new LogFindAuditEventsRequestType();
-            logReqMsg.setMessage(message);
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            auditMsg = FindAuditEventsTransforms.transformFindAuditEventsReq2AuditMsg(logReqMsg);
+        LogFindAuditEventsRequestType logReqMsg = new LogFindAuditEventsRequestType();
+        logReqMsg.setMessage(message);
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        auditMsg = FindAuditEventsTransforms.transformFindAuditEventsReq2AuditMsg(logReqMsg);
         LOG.debug("Exiting AuditRepositoryLogger.logFindAuditEvents(...)");
         return auditMsg;
     }
@@ -513,14 +519,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logFindAuditEventsResult(FindAuditEventsResponseMessageType message, String direction,
-            String _interface) {
+        String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logFindAuditEventsResult(...)");
         LogEventRequestType auditMsg = null;
-            LogFindAuditEventsResultRequestType logReqMsg = new LogFindAuditEventsResultRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            LOG.warn("logFindAuditEventsResult method is not implemented");
+        LogFindAuditEventsResultRequestType logReqMsg = new LogFindAuditEventsResultRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        LOG.warn("logFindAuditEventsResult method is not implemented");
         LOG.debug("Exiting AuditRepositoryLogger.logFindAuditEventsResult(...)");
         return auditMsg;
     }
@@ -536,13 +542,13 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
     public LogEventRequestType logNhinSubscribeRequest(SubscribeRequestType message, String direction, String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logNhinSubscribeRequest(...)");
         LogEventRequestType auditMsg = null;
-            SubscribeTransforms transformLib = new SubscribeTransforms();
-            LogNhinSubscribeRequestType logReqMsg = new LogNhinSubscribeRequestType();
-            logReqMsg.setMessage(message);
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
+        SubscribeTransforms transformLib = new SubscribeTransforms();
+        LogNhinSubscribeRequestType logReqMsg = new LogNhinSubscribeRequestType();
+        logReqMsg.setMessage(message);
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
 
-            auditMsg = transformLib.transformNhinSubscribeRequestToAuditMessage(logReqMsg);
+        auditMsg = transformLib.transformNhinSubscribeRequestToAuditMessage(logReqMsg);
         LOG.debug("Exiting AuditRepositoryLogger.logNhinSubscribeRequest(...)");
         return auditMsg;
     }
@@ -558,12 +564,12 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
     public LogEventRequestType logNhinNotifyRequest(NotifyRequestType message, String direction, String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logNhinNotifyRequest(...)");
         LogEventRequestType auditMsg = null;
-            LogNhinNotifyRequestType logReqMsg = new LogNhinNotifyRequestType();
-            logReqMsg.setMessage(message);
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
+        LogNhinNotifyRequestType logReqMsg = new LogNhinNotifyRequestType();
+        logReqMsg.setMessage(message);
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
 
-            auditMsg = transformLib.transformNhinNotifyRequestToAuditMessage(logReqMsg);
+        auditMsg = transformLib.transformNhinNotifyRequestToAuditMessage(logReqMsg);
         LOG.debug("Exiting AuditRepositoryLogger.logNhinNotifyRequest(...)");
         return auditMsg;
     }
@@ -577,15 +583,15 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logNhinUnsubscribeRequest(UnsubscribeRequestType message, String direction,
-            String _interface) {
+        String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logNhinUnsubscribeRequest(...)");
         LogEventRequestType auditMsg = null;
-            UnsubscribeTransforms transformLib = new UnsubscribeTransforms();
-            LogNhinUnsubscribeRequestType logReqMsg = new LogNhinUnsubscribeRequestType();
-            logReqMsg.setMessage(message);
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            auditMsg = transformLib.transformNhinUnsubscribeRequestToAuditMessage(logReqMsg);
+        UnsubscribeTransforms transformLib = new UnsubscribeTransforms();
+        LogNhinUnsubscribeRequestType logReqMsg = new LogNhinUnsubscribeRequestType();
+        logReqMsg.setMessage(message);
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        auditMsg = transformLib.transformNhinUnsubscribeRequestToAuditMessage(logReqMsg);
         LOG.debug("Exiting AuditRepositoryLogger.logNhinUnsubscribeRequest(...)");
         return auditMsg;
     }
@@ -599,15 +605,15 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logUnsubscribeResponse(UnsubscribeResponseMessageType message, String direction,
-            String _interface) {
+        String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logUnsubscribeResponse(...)");
         LogEventRequestType auditMsg = null;
-            LogUnsubscribeResponseType logReqMsg = new LogUnsubscribeResponseType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            UnsubscribeTransforms transformLib = new UnsubscribeTransforms();
-            auditMsg = transformLib.transformUnsubscribeResponseToGenericAudit(logReqMsg);
+        LogUnsubscribeResponseType logReqMsg = new LogUnsubscribeResponseType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        UnsubscribeTransforms transformLib = new UnsubscribeTransforms();
+        auditMsg = transformLib.transformUnsubscribeResponseToGenericAudit(logReqMsg);
         LOG.debug("Exiting AuditRepositoryLogger.logUnsubscribeResponse(...)");
         return auditMsg;
     }
@@ -621,15 +627,15 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logSubscribeResponse(SubscribeResponseMessageType message, String direction,
-            String _interface) {
+        String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logSubscribeResponse(...)");
         LogEventRequestType auditMsg = null;
-            LogSubscribeResponseType logReqMsg = new LogSubscribeResponseType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            SubscribeTransforms transformLib = new SubscribeTransforms();
-            auditMsg = transformLib.transformSubscribeResponseToAuditMessage(logReqMsg);
+        LogSubscribeResponseType logReqMsg = new LogSubscribeResponseType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        SubscribeTransforms transformLib = new SubscribeTransforms();
+        auditMsg = transformLib.transformSubscribeResponseToAuditMessage(logReqMsg);
         LOG.debug("Exiting AuditRepositoryLogger.logSubscribeResponse(...)");
         return auditMsg;
     }
@@ -643,14 +649,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityDocSubscribeRequest(EntityDocumentSubscribeRequestMessageType message,
-            String direction, String _interface) {
+        String direction, String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityDocSubscribeRequest(...)");
         LogEventRequestType auditMsg = null;
-            LogEntityDocumentSubscribeRequestType logReqMsg = new LogEntityDocumentSubscribeRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            LOG.warn("logEntityDocSubscribeRequest method is not implemented");
+        LogEntityDocumentSubscribeRequestType logReqMsg = new LogEntityDocumentSubscribeRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        LOG.warn("logEntityDocSubscribeRequest method is not implemented");
         LOG.debug("Exiting AuditRepositoryLogger.logEntityDocSubscribeRequest(...)");
         return auditMsg;
     }
@@ -664,14 +670,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityCdcSubscribeRequest(EntityCdcSubscribeRequestMessageType message,
-            String direction, String _interface) {
+        String direction, String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityCdcSubscribeRequest(...)");
         LogEventRequestType auditMsg = null;
-            LogEntityCdcSubscribeRequestType logReqMsg = new LogEntityCdcSubscribeRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            LOG.warn("logEntityCdcSubscribeRequest method is not implemented");
+        LogEntityCdcSubscribeRequestType logReqMsg = new LogEntityCdcSubscribeRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        LOG.warn("logEntityCdcSubscribeRequest method is not implemented");
         LOG.debug("Exiting AuditRepositoryLogger.logEntityCdcSubscribeRequest(...)");
         return auditMsg;
     }
@@ -685,14 +691,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityDocNotifyRequest(EntityDocumentNotifyRequestMessageType message,
-            String direction, String _interface) {
+        String direction, String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityDocNotifyRequest(...)");
         LogEventRequestType auditMsg = null;
-            LogEntityDocumentNotifyRequestType logReqMsg = new LogEntityDocumentNotifyRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            LOG.warn("logEntityDocNotifyRequest method is not implemented");
+        LogEntityDocumentNotifyRequestType logReqMsg = new LogEntityDocumentNotifyRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        LOG.warn("logEntityDocNotifyRequest method is not implemented");
         LOG.debug("Exiting AuditRepositoryLogger.logEntityDocNotifyRequest(...)");
         return auditMsg;
     }
@@ -706,14 +712,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityCdcNotifyRequest(EntityCdcNotifyRequestMessageType message, String direction,
-            String _interface) {
+        String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityCdcNotifyRequest(...)");
         LogEventRequestType auditMsg = null;
-            LogEntityCdcNotifyRequestType logReqMsg = new LogEntityCdcNotifyRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            LOG.warn("logEntityCdcNotifyRequest method is not implemented");
+        LogEntityCdcNotifyRequestType logReqMsg = new LogEntityCdcNotifyRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        LOG.warn("logEntityCdcNotifyRequest method is not implemented");
         LOG.debug("Exiting AuditRepositoryLogger.logEntityCdcNotifyRequest(...)");
         return auditMsg;
     }
@@ -727,14 +733,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityNotifyResponse(EntityNotifyResponseMessageType message, String direction,
-            String _interface) {
+        String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityNotifyResponse(...)");
         LogEventRequestType auditMsg = null;
-            LogEntityNotifyResponseType logReqMsg = new LogEntityNotifyResponseType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
-            LOG.warn("logEntityNotifyRespRequest method is not implemented");
+        LogEntityNotifyResponseType logReqMsg = new LogEntityNotifyResponseType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
+        LOG.warn("logEntityNotifyRespRequest method is not implemented");
         LOG.debug("Exiting AuditRepositoryLogger.logEntityNotifyResponse(...)");
         return auditMsg;
     }
@@ -748,7 +754,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return
      */
     public LogEventRequestType logXDRReq(ProvideAndRegisterDocumentSetRequestType message, AssertionType assertion,
-            NhinTargetSystemType target, String direction) {
+        NhinTargetSystemType target, String direction) {
         LOG.debug("Entering AuditRepositoryLogger.logNhinXDRReq(...)");
         if (message == null) {
             LOG.error("Message is null");
@@ -756,7 +762,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
         }
         XDRTransforms auditTransformer = new XDRTransforms();
         LogEventRequestType auditMsg = auditTransformer.transformRequestToAuditMsg(message, assertion, target, direction,
-                NhincConstants.AUDIT_LOG_NHIN_INTERFACE);
+            NhincConstants.AUDIT_LOG_NHIN_INTERFACE);
         LOG.debug("Exiting AuditRepositoryLogger.logNhinXDRReq(...)");
         return auditMsg;
     }
@@ -770,7 +776,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return
      */
     public LogEventRequestType logAdapterXDRReq(ProvideAndRegisterDocumentSetRequestType message, AssertionType assertion,
-            String direction) {
+        String direction) {
         LOG.debug("Entering AuditRepositoryLogger.logAdapterXDRReq(...)");
         LogEventRequestType auditMsg = null;
 
@@ -781,7 +787,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
 
         XDRTransforms auditTransformer = new XDRTransforms();
         auditMsg = auditTransformer.transformRequestToAuditMsg(message, assertion, null, direction,
-                NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE);
+            NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE);
 
         LOG.debug("Exiting AuditRepositoryLogger.logAdapterXDRReq(...)");
         return auditMsg;
@@ -796,12 +802,12 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return
      */
     public LogEventRequestType logXDRReq(
-            gov.hhs.fha.nhinc.common.nhinccommonproxy.RespondingGatewayProvideAndRegisterDocumentSetSecuredRequestType message,
-            AssertionType assertion, NhinTargetSystemType target, String direction) {
+        gov.hhs.fha.nhinc.common.nhinccommonproxy.RespondingGatewayProvideAndRegisterDocumentSetSecuredRequestType message,
+        AssertionType assertion, NhinTargetSystemType target, String direction) {
         LOG.debug("Entering AuditRepositoryLogger.logXDRReq(...)");
         XDRTransforms auditTransformer = new XDRTransforms();
         LogEventRequestType auditMsg = auditTransformer.transformRequestToAuditMsg(message, assertion, target,
-                direction, NhincConstants.AUDIT_LOG_NHIN_INTERFACE);
+            direction, NhincConstants.AUDIT_LOG_NHIN_INTERFACE);
         LOG.debug("Exiting AuditRepositoryLogger.logNhinXDRReq(...)");
         return auditMsg;
     }
@@ -815,12 +821,12 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return
      */
     public LogEventRequestType logNhinXDRResponse(RegistryResponseType message, AssertionType assertion,
-            NhinTargetSystemType target, String direction, boolean isRequesting) {
+        NhinTargetSystemType target, String direction, boolean isRequesting) {
         LOG.debug("Entering AuditRepositoryLogger.logNhinXDRResponse(...)");
         LogEventRequestType auditMsg = null;
-            XDRTransforms auditTransformer = new XDRTransforms();
-            auditMsg = auditTransformer.transformResponseToAuditMsg(message, assertion, target, direction,
-                    NhincConstants.AUDIT_LOG_NHIN_INTERFACE, isRequesting);
+        XDRTransforms auditTransformer = new XDRTransforms();
+        auditMsg = auditTransformer.transformResponseToAuditMsg(message, assertion, target, direction,
+            NhincConstants.AUDIT_LOG_NHIN_INTERFACE, isRequesting);
         LOG.debug("Exiting AuditRepositoryLogger.logNhinXDRResponse(...)");
         return auditMsg;
     }
@@ -834,21 +840,22 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityUnsubscribeRequest(EntityUnsubscribeRequestMessageType message,
-            String direction, String _interface) {
+        String direction, String _interface) {
         LOG.debug("Entering AuditRepositoryLogger.logEntityUnsubscribeRequest(...)");
         LogEventRequestType auditMsg = null;
-            LogEntityUnsubscribeRequestType logReqMsg = new LogEntityUnsubscribeRequestType();
-            logReqMsg.setDirection(direction);
-            logReqMsg.setInterface(_interface);
-            logReqMsg.setMessage(message);
+        LogEntityUnsubscribeRequestType logReqMsg = new LogEntityUnsubscribeRequestType();
+        logReqMsg.setDirection(direction);
+        logReqMsg.setInterface(_interface);
+        logReqMsg.setMessage(message);
 
-            LOG.warn("logEntityUnsubscribeRequest method is not implemented");
+        LOG.warn("logEntityUnsubscribeRequest method is not implemented");
         LOG.debug("Exiting AuditRepositoryLogger.logEntityUnsubscribeRequest(...)");
         return auditMsg;
     }
 
     /**
      * Log an XDR Acknowledgement at the Nhin interface.
+     *
      * @param acknowledgement The XDR Acknowledgement to be audited
      * @param assertion The assertion to be audited
      * @param direction The direction this message is going (Inbound or Outbound)
@@ -856,10 +863,10 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return
      */
     public LogEventRequestType logAcknowledgement(XDRAcknowledgementType acknowledgement, AssertionType assertion,
-            NhinTargetSystemType target, String direction, String action) {
+        NhinTargetSystemType target, String direction, String action) {
         LOG.debug("Entering AuditRepositoryLogger.logAcknowledgement(...)");
         LogEventRequestType auditMsg = xdrAuditTransformer.transformAcknowledgementToAuditMsg(acknowledgement,
-                assertion, target, direction, NhincConstants.AUDIT_LOG_NHIN_INTERFACE, action);
+            assertion, target, direction, NhincConstants.AUDIT_LOG_NHIN_INTERFACE, action);
 
         LOG.debug("Exiting AuditRepositoryLogger.logAcknowledgement(...)");
         return auditMsg;
@@ -867,6 +874,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
 
     /**
      * Log an XDR Acknowledgement at the Adapter interface.
+     *
      * @param acknowledgement The XDR Acknowledgement to be audited
      * @param assertion The assertion to be audited
      * @param direction The direction this message is going (Inbound or Outbound)
@@ -874,16 +882,17 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return
      */
     public LogEventRequestType logAdapterAcknowledgement(XDRAcknowledgementType acknowledgement, AssertionType assertion,
-            String direction, String action) {
+        String direction, String action) {
         LOG.debug("Entering AuditRepositoryLogger.logAdapterAcknowledgement(...)");
         LogEventRequestType auditMsg = xdrAuditTransformer.transformAcknowledgementToAuditMsg(acknowledgement,
-                assertion, null, direction, NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE, action);
+            assertion, null, direction, NhincConstants.AUDIT_LOG_ADAPTER_INTERFACE, action);
         LOG.debug("Exiting AuditRepositoryLogger.logAdapterAcknowledgement(...)");
         return auditMsg;
     }
 
     /**
      * Log an XDR Acknowledgement at the Entity interface.
+     *
      * @param acknowledgement The XDR Acknowledgement to be audited
      * @param assertion The assertion to be audited
      * @param direction The direction this message is going (Inbound or Outbound)
@@ -891,10 +900,10 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return
      */
     public LogEventRequestType logEntityAcknowledgement(XDRAcknowledgementType acknowledgement,
-            AssertionType assertion, String direction, String action) {
+        AssertionType assertion, String direction, String action) {
         LOG.debug("Entering AuditRepositoryLogger.logAcknowledgement(...)");
         LogEventRequestType auditMsg = xdrAuditTransformer.transformAcknowledgementToAuditMsg(acknowledgement,
-                assertion, null, direction, NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, action);
+            assertion, null, direction, NhincConstants.AUDIT_LOG_ENTITY_INTERFACE, action);
         LOG.debug("Exiting AuditRepositoryLogger.logAcknowledgement(...)");
         return auditMsg;
     }
@@ -908,17 +917,17 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return A generic audit log message that can be passed to the Audit Repository
      */
     public LogEventRequestType logEntityAdminDist(
-            gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewaySendAlertMessageType message,
-            AssertionType assertion, String direction) {
-    	LOG.trace("Entering AuditRepositoryLogger.logEntityPatientDiscResp(...)");
+        gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewaySendAlertMessageType message,
+        AssertionType assertion, String direction) {
+        LOG.trace("Entering AuditRepositoryLogger.logEntityPatientDiscResp(...)");
 
-    	LogEventRequestType auditMsg = null;
-        try{
-        	EDXLDistribution body = message.getEDXLDistribution();
-        	auditMsg = adAuditTransformer.transformEDXLDistributionRequestToAuditMsg(body, assertion, direction,
-                    NhincConstants.AUDIT_LOG_ENTITY_INTERFACE);
-        }catch (NullPointerException ex){
-        	LOG.error("The Incoming Send Alert message was Null", ex);
+        LogEventRequestType auditMsg = null;
+        try {
+            EDXLDistribution body = message.getEDXLDistribution();
+            auditMsg = adAuditTransformer.transformEDXLDistributionRequestToAuditMsg(body, assertion, direction,
+                NhincConstants.AUDIT_LOG_ENTITY_INTERFACE);
+        } catch (NullPointerException ex) {
+            LOG.error("The Incoming Send Alert message was Null", ex);
         }
 
         LOG.trace("Exiting AuditRepositoryLogger.logEntityPatientDiscResp(...)");
@@ -927,6 +936,7 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
 
     /**
      * Create a generic Audit Log for an Admin Distribution message.
+     *
      * @param message The Admin Distribution message to be audited
      * @param assertion The assertion to be audited
      * @param target
@@ -935,13 +945,14 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return
      */
     public LogEventRequestType logNhincAdminDist(EDXLDistribution message, AssertionType assertion,
-            NhinTargetSystemType target, String direction, String _interface) {
+        NhinTargetSystemType target, String direction, String _interface) {
 
-    	return adAuditTransformer.transformEDXLDistributionRequestToAuditMsg(message, assertion, target, direction, _interface);
+        return adAuditTransformer.transformEDXLDistributionRequestToAuditMsg(message, assertion, target, direction, _interface);
     }
 
     /**
      * Create a generic Audit Log for an Admin Distribution message.
+     *
      * @param message The Admin Distribution message to be audited
      * @param assertion The assertion to be audited
      * @param direction The direction this message is going (Inbound or Outbound)
@@ -949,9 +960,80 @@ public class AuditRepositoryLogger implements AuditRepositoryDocumentRetrieveLog
      * @return
      */
     public LogEventRequestType logNhincAdminDist(EDXLDistribution message, AssertionType assertion, String direction,
-            String _interface) {
+        String _interface) {
 
-    	return adAuditTransformer.transformEDXLDistributionRequestToAuditMsg(message, assertion, direction, _interface);
+        return adAuditTransformer.transformEDXLDistributionRequestToAuditMsg(message, assertion, direction, _interface);
+    }
+
+    /**
+     *
+     * @param message
+     * @param assertion
+     * @param target
+     * @param direction
+     * @param isRequesting
+     * @param webContextProperties
+     * @param serviceName
+     */
+    public void auditNhinCoreX12BatchMessage(Object message, AssertionType assertion, NhinTargetSystemType target,
+        String direction, boolean isRequesting, Properties webContextProperties, String serviceName) {
+
+        if (getAuditRepositoryProxyObjectFactory().isNoopImplementationUsed()) {
+            return;
+        }
+        LOG.trace("---Begin AuditRepositoryLogger.auditNhinCoreX12BatchRequest()---");
+        LogEventRequestType auditLogMsg = coreX12AuditDataTransform.transformX12MsgToAuditMsg(message, assertion, target, direction, NhincConstants.AUDIT_LOG_NHIN_INTERFACE, isRequesting, webContextProperties, serviceName);
+        if (auditLogMsg != null && auditLogMsg.getAuditMessage() != null) {
+            audit(auditLogMsg, assertion);
+        } else {
+            LOG.error("Core X12 Nhin Batch Request auditLogMsg is null");
+        }
+        LOG.trace("---End AuditRepositoryLogger.auditNhinCoreX12BatchRequest()---");
+    }
+
+    /**
+     *
+     * @param message
+     * @param assertion
+     * @param target
+     * @param direction
+     * @param isRequesting
+     * @param webContextProperties
+     * @param serviceName
+     */
+    public void auditNhinCoreX12RealtimeMessage(Object message, AssertionType assertion, NhinTargetSystemType target,
+        String direction, boolean isRequesting, Properties webContextProperties, String serviceName) {
+        if (getAuditRepositoryProxyObjectFactory().isNoopImplementationUsed()) {
+            return;
+        }
+        LOG.info("Inside the method AuditRepositoryLogger.auditNhinCoreX12RealtimeMessage (START)-->TimeStamp:" + (new Date()));
+        LOG.trace("---Begin AuditRepositoryLogger.auditNhinCoreX12RealtimeMessage()---");
+        // Set up the audit logging request message
+        LogEventRequestType auditLogMsg = coreX12AuditDataTransform.transformX12MsgToAuditMsg(message, assertion, target, direction, NhincConstants.AUDIT_LOG_NHIN_INTERFACE, isRequesting, webContextProperties, serviceName);
+        if (auditLogMsg != null && auditLogMsg.getAuditMessage() != null) {
+            audit(auditLogMsg, assertion);
+        } else {
+            LOG.error("Core X12 Realtime Request auditLogMsg is null");
+        }
+        LOG.trace("---End AuditRepositoryLogger.auditNhinCoreX12RealtimeMessage()---");
+        LOG.info("Inside the method AuditRepositoryLogger.auditNhinCoreX12RealtimeMessage (END)-->TimeStamp:" + (new Date()));
+    }
+
+    private AuditRepositoryProxyObjectFactory getAuditRepositoryProxyObjectFactory() {
+        //TODO: AuditRepositoryProxyObjectFactory potential candidate for single ton
+        return (new AuditRepositoryProxyObjectFactory());
+    }
+
+    /**
+     * Submits a generic Audit Log message to the Audit Log Repository.
+     *
+     * @param auditLogMsg The generic audit log to be audited
+     * @param assertion The assertion to be audited
+     * @return
+     */
+    private AcknowledgementType audit(LogEventRequestType auditLogMsg, AssertionType assertion) {
+        AuditRepositoryProxy proxy = getAuditRepositoryProxyObjectFactory().getAuditRepositoryProxy();
+        return proxy.auditLog(auditLogMsg, assertion);
     }
 
 }
